move connection state
