+++ src/components/connection-state/ConnectionState.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled, { withTheme } from 'styled-components';
 import PropTypes from 'prop-types';
-import ConnectionIndicator, { getColor } from '../ConnectionIndicator';
+import ConnectionIndicator, { getColor } from './ConnectionIndicator';
 
 const propTypes = {
   connectionState: PropTypes.string.isRequired
